Change price and area data types Remove sub-layout pages and use the global layout Adjust the menus to fit the root layout Finish add building UI

+++ HammadBroker.Model/DTO/BuildingForDetails.cs
@@ -11,7 +11,7 @@ public class BuildingForDetails : BuildingForList
 	[Display(Name = "الحمامات")]
 	public byte? Bathrooms { get; set; }
 	[Display(Name = "المساحة")]
-	public decimal? Area { get; set; }
+	public long? Area { get; set; }
 	[Display(Name = "العنوان")]
 	public string Address { get; set; }
 	[Display(Name = "الحي")]
